UNIONVMS-597 Filter table of movements (R0142 - a)
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Reporting/APP@5421 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/reporting/service/entities/VmsPositionFilter.java
@@ -19,10 +19,10 @@ import java.util.List;
 public class VmsPositionFilter extends Filter  {
 
     @Column(name = "MIN_SPEED")
-    private Long minimumSpeed;
+    private Float minimumSpeed;
 
     @Column(name = "MAX_SPEED")
-    private Long maximumSpeed;
+    private Float maximumSpeed;
 
     @Column(name = "MOV_TYPE")
     private MovementTypeType movementType;
@@ -59,11 +59,11 @@ public class VmsPositionFilter extends Filter  {
         this.setMinimumSpeed(incoming.getMinimumSpeed());
     }
 
-    public Long getMinimumSpeed() {
+    public Float getMinimumSpeed() {
         return minimumSpeed;
     }
 
-    public void setMinimumSpeed(Long minimumSpeed) {
+    public void setMinimumSpeed(Float minimumSpeed) {
         this.minimumSpeed = minimumSpeed;
     }
 
@@ -83,11 +83,11 @@ public class VmsPositionFilter extends Filter  {
         this.movementActivity = movementActivity;
     }
 
-    public Long getMaximumSpeed() {
+    public Float getMaximumSpeed() {
         return maximumSpeed;
     }
 
-    public void setMaximumSpeed(Long maximumSpeed) {
+    public void setMaximumSpeed(Float maximumSpeed) {
         this.maximumSpeed = maximumSpeed;
     }
